feat: separated fetch function from service to use it generally
