+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/finance/FinanceService.java
@@ -1,55 +1,13 @@
 package com.ssafy.shinhanflow.finance;
 
-import static com.ssafy.shinhanflow.config.error.ErrorCode.INTERNAL_SERVER_ERROR;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
-import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.util.UriComponentsBuilder;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ssafy.shinhanflow.config.error.exception.BusinessBaseException;
-import com.ssafy.shinhanflow.finance.dto.FinanceApiRequestDto;
-import com.ssafy.shinhanflow.finance.dto.FinanceApiResponseDto;
-import com.ssafy.shinhanflow.finance.dto.MemberRequestDto;
-import com.ssafy.shinhanflow.finance.dto.MemberResponseDto;
+import com.ssafy.shinhanflow.financeapi.FinanceApiFetcher;
 
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Service
 public class FinanceService {
-	@Value("${finance-api.key}")
-	private String apiKey;
-
-	@Value("${finance-api.base-url}")
-	private String baseUrl;
-
-	private final WebClient webClient;
-	private final ObjectMapper objectMapper;
-
-	private <T extends FinanceApiResponseDto> T fetch(String urlPath, FinanceApiRequestDto financeApiRequestDto,
-		Class<T> responseType) {
-		financeApiRequestDto.setApiKey(apiKey);
-		T response;
-		try {
-			response = webClient.post()
-				.uri(UriComponentsBuilder.fromHttpUrl(baseUrl).path(urlPath).toUriString())
-				.contentType(MediaType.APPLICATION_JSON)
-				.bodyValue(objectMapper.writeValueAsString(financeApiRequestDto))
-				.retrieve()
-				.bodyToMono(responseType)
-				.block();
-		} catch (JsonProcessingException e) {
-			throw new BusinessBaseException("Failed to serialize request", INTERNAL_SERVER_ERROR);
-		}
-		return response;
-	}
-
-	public MemberResponseDto createMember(MemberRequestDto memberRequestDto) {
-		return fetch("member", memberRequestDto, MemberResponseDto.class);
-	}
-
+	FinanceApiFetcher financeApiFetcher;
 }
